+++ webstore/src/main/java/com/packt/webstore/service/StudentService.java
@@ -8,7 +8,7 @@ import com.packt.webstore.domain.Supervisor;
 public interface StudentService {
 public Student getStudentById(int id) ;
 public List<Student> getStudents();
-public abstract Supervisor getStudentSupervisor(int studentId);
+public abstract Supervisor getStudentSupervisor(long studentId);
 public boolean registerStudent(int studentId, String studentName, String studentEmail, String department);
 public Student getStudentByEmail(String userEmail);
 public boolean updateStudent(int studentId, String cellPhone, String password);
